requirements.txt
aiofiles==24.1.0
aiohttp==3.9.1
aiosignal==1.3.2
anyio==4.2.0
attrs==23.1.0
azure-core==1.30.1
azure-identity==1.17.1
azure-storage-blob==12.24.1
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
deepgram-sdk==2.12.0
Flask==2.2.2
Flask-Cors==5.0.0
frozenlist==1.5.0
gunicorn==20.1.0
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.5
lxml==5.3.0
MarkupSafe==3.0.2
msal==1.31.1
msal-extensions==1.2.0
multidict==6.1.0
nest-asyncio==1.5.6
openai==0.28.0
portalocker==2.10.1
propcache==0.2.1
pycparser==2.22
PyJWT==2.10.1
python-docx==1.1.2
python-dotenv==1.0.1
requests==2.31.0
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.9.0
urllib3==2.3.0
websockets==14.2
Werkzeug==2.2.3
yarl==1.18.3
#tiktoken==0.5.1

imageio[ffmpeg]==2.37.0  # <- imageio の拡張機能を含む方を採用
ffmpeg-python==0.2.0
future==1.0.0
imageio-ffmpeg==0.6.0
numpy
pillow==11.1.0
psutil==6.1.1
pydub
waitress>=2.1.0
azure-functions
azure-storage-queue>=12.5.0



